Refactored lexing to be a child class of a more general token stream
